feat: rename function/file, add switch statement
To not duplicate a function with almost the same structure I added a layer of flexibility by adding a switch. The function got renamed to getContent and takes now a property of kind which allows to fetch static_content as well as projects content from the database, depending of the type of kind which is provided.
